OsintAraci2/requirements.txt
requests>=2.31.0
urllib3>=2.0.7
whois>=0.9.27
dnspython>=2.4.2
scapy>=2.5.0
Pillow>=10.0.0
scrapy>=2.11.1
itemadapter>=0.8.0
googlesearch-python>=1.2.3
ttkthemes>=3.2.2
ttkwidgets>=0.13.0
